requirements.txt
# =========================
# Core Framework
# =========================
fastapi==0.115.6
uvicorn[standard]==0.30.6

# =========================
# Database & ORM
# =========================
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9

# =========================
# Async & Performance
# =========================
asyncpg==0.29.0
httpx==0.27.2
anyio==4.6.2

# =========================
# Authentication & Security
# =========================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12

# =========================
# Validation & Settings
# =========================
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# =========================
# Caching & Rate Limiting
# =========================
redis==5.2.0
slowapi==0.1.9

# =========================
# Background Tasks & Workers
# =========================
celery==5.4.0
kombu==5.4.2

# =========================
# Payments
# =========================
stripe==11.6.0

# =========================
# Media & Storage
# =========================
boto3==1.35.54
botocore==1.35.54

# =========================
# Logging & Monitoring
# =========================
loguru==0.7.2
sentry-sdk==2.19.0

# =========================
# Utilities
# =========================
python-dotenv==1.0.1
tenacity==9.0.0
uuid6==2024.7.10

# =========================
# Testing
# =========================
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# =========================
# Development & Formatting
# =========================
black==24.10.0
isort==5.13.2
mypy==1.13.0
